third_party/blink/web_tests/platform/linux-chrome/external/wpt/dom/ranges/Range-surroundContents-expected.txt

This is a testharness.js-based test.
Found 40 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] 6,0: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node paras[0]
  Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node.
[FAIL] 6,0: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node paras[0]
  assert_true: First differing node: expected Text node "Ijklmnop\n", got Text node "A虉b虉c虉d虉e虉f虉g虉h虉\n" [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,1: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node paras[0].firstChild
  assert_throws_dom: A HIERARCHY_REQUEST_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException HIERARCHY_REQUEST_ERR: property "code" is equal to 11, expected 3
[FAIL] 6,1: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node paras[0].firstChild
  assert_true: First differing node: expected Element node <p id="b" style="display:none">Ijklmnop\n</p>, got Text node "A虉b虉c虉d虉e虉f虉g虉h虉\n" [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,2: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node paras[1].firstChild
  assert_throws_dom: A HIERARCHY_REQUEST_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException HIERARCHY_REQUEST_ERR: property "code" is equal to 11, expected 3
[FAIL] 6,2: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node paras[1].firstChild
  assert_true: First differing node: expected Element node <p id="c">Qrstuvwx</p>, got Text node "Ijklmnop\n" [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,3: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node paras[5].firstChild
  assert_throws_dom: A HIERARCHY_REQUEST_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException HIERARCHY_REQUEST_ERR: property "code" is equal to 11, expected 3
[FAIL] 6,3: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node paras[5].firstChild
  assert_true: First differing node: expected Node object of unknown type, got Node object of unknown type [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,4: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node foreignPara1
  Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node.
[FAIL] 6,4: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node foreignPara1
  assert_true: First differing node: expected Node object of unknown type, got Node object of unknown type [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,5: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node foreignPara1.firstChild
  assert_throws_dom: A HIERARCHY_REQUEST_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException HIERARCHY_REQUEST_ERR: property "code" is equal to 11, expected 3
[FAIL] 6,5: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node foreignPara1.firstChild
  assert_true: First differing node: expected Node object of unknown type, got Node object of unknown type [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,6: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node detachedPara1
  Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node.
[FAIL] 6,6: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node detachedPara1
  assert_true: First differing node: expected Node object of unknown type, got Node object of unknown type [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,7: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node detachedPara1.firstChild
  assert_throws_dom: A HIERARCHY_REQUEST_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException HIERARCHY_REQUEST_ERR: property "code" is equal to 11, expected 3
[FAIL] 6,7: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node detachedPara1.firstChild
  assert_true: First differing node: expected Node object of unknown type, got Node object of unknown type [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,8: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node document
  assert_throws_dom: A INVALID_NODE_TYPE_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException INVALID_NODE_TYPE_ERR: property "code" is equal to 11, expected 24
[FAIL] 6,9: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node detachedDiv
  Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node.
[FAIL] 6,9: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node detachedDiv
  assert_true: First differing node: expected Node object of unknown type, got Node object of unknown type [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,10: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node foreignDoc
  assert_throws_dom: A INVALID_NODE_TYPE_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException INVALID_NODE_TYPE_ERR: property "code" is equal to 11, expected 24
[FAIL] 6,11: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node foreignPara2
  Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node.
[FAIL] 6,11: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node foreignPara2
  assert_true: First differing node: expected Node object of unknown type, got Node object of unknown type [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,12: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node xmlDoc
  assert_throws_dom: A INVALID_NODE_TYPE_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException INVALID_NODE_TYPE_ERR: property "code" is equal to 11, expected 24
[FAIL] 6,13: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node xmlElement
  Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node.
[FAIL] 6,13: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node xmlElement
  assert_true: First differing node: expected Node object of unknown type, got Node object of unknown type [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,14: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node detachedTextNode
  assert_throws_dom: A HIERARCHY_REQUEST_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException HIERARCHY_REQUEST_ERR: property "code" is equal to 11, expected 3
[FAIL] 6,14: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node detachedTextNode
  assert_true: First differing node: expected Node object of unknown type, got Node object of unknown type [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,15: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node foreignTextNode
  assert_throws_dom: A HIERARCHY_REQUEST_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException HIERARCHY_REQUEST_ERR: property "code" is equal to 11, expected 3
[FAIL] 6,15: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node foreignTextNode
  assert_true: First differing node: expected Node object of unknown type, got Node object of unknown type [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,16: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node processingInstruction
  assert_throws_dom: A HIERARCHY_REQUEST_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException HIERARCHY_REQUEST_ERR: property "code" is equal to 11, expected 3
[FAIL] 6,16: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node processingInstruction
  assert_true: First differing node: expected Node object of unknown type, got Node object of unknown type [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,17: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node detachedProcessingInstruction
  assert_throws_dom: A HIERARCHY_REQUEST_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException HIERARCHY_REQUEST_ERR: property "code" is equal to 11, expected 3
[FAIL] 6,17: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node detachedProcessingInstruction
  assert_true: First differing node: expected Node object of unknown type, got Node object of unknown type [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,18: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node comment
  assert_throws_dom: A HIERARCHY_REQUEST_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException HIERARCHY_REQUEST_ERR: property "code" is equal to 11, expected 3
[FAIL] 6,18: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node comment
  assert_true: First differing node: expected Node object of unknown type, got Node object of unknown type [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,19: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node detachedComment
  assert_throws_dom: A HIERARCHY_REQUEST_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException HIERARCHY_REQUEST_ERR: property "code" is equal to 11, expected 3
[FAIL] 6,19: resulting range position for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node detachedComment
  assert_true: First differing node: expected Node object of unknown type, got Node object of unknown type [Actual and expected mismatch for range's tree root.  ] expected true got false
[FAIL] 6,20: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node docfrag
  assert_throws_dom: A INVALID_NODE_TYPE_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException INVALID_NODE_TYPE_ERR: property "code" is equal to 11, expected 24
[FAIL] 6,21: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node doctype
  assert_throws_dom: A INVALID_NODE_TYPE_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException INVALID_NODE_TYPE_ERR: property "code" is equal to 11, expected 24
[FAIL] 6,22: resulting DOM for range [paras[5].firstChild, 2, paras[5].lastChild, 4], node foreignDoctype
  assert_throws_dom: A INVALID_NODE_TYPE_ERR must be thrown in this case function "function() {\n        try {\n          actualRange.surroundContents(actualNode);\n        } catch (e) {\n          if (expectedDoctype != expectedIframe.contentDocument.firstChild) {\n            expectedIframe.contentDocument.insertBefore(expectedDoctype, expectedIframe.contentDocument.firstChild);\n          }\n          if (actualDoctype != actualIframe.contentDocument.firstChild) {\n            actualIframe.contentDocument.insertBefore(actualDoctype, actualIframe.contentDocument.firstChild);\n          }\n          throw e;\n        }\n      }" threw object "InvalidStateError: Failed to execute 'surroundContents' on 'Range': The Range has partially selected a non-Text node." that is not a DOMException INVALID_NODE_TYPE_ERR: property "code" is equal to 11, expected 24
Harness: the test ran to completion.
